#4: refactored - divided by modules
